+++ CHANGELOG.md
@@ -28,7 +28,21 @@ This document details all notable changes to the WSU HRS Child Theme. Uses [Sema
 - Set up grid fallbacks, open #31.
 - Set up no-js handling for the search form and menu, open #34.
 - Get Babel processing working, open #35.
-- Determine if we really need the `center-middle` utility class. If yes, look into using a flex or grid with `Xvh` and `Xvh` column and row size to center instead.
+
+## 0.15.0 (2018-07-06)
+
+### Changed
+
+- Update archive pages (`home.php`, `archive.php`, and `taxonomy-hrs_unit.php`) to display the most recent 10 posts, include pagination, and only use the special feature layout on the first results page.
+- Update page header.
+
+### Added
+
+- Base styles for the posts home page.
+- Method to display a gallery of taxonomy term titles.
+- Method to retrieve recent posts categorized as "reminders."
+- Icons for bookmarks and events.
+- A `home.php` template to handle display and layout of the home posts page.
 
 ## 0.14.0 (2018-06-27)